Move get_quotas into services
